Update Camera.js
if the object has the function isVisable() then it will test if it passes true of false to know if the object should be rendered. if the object lacks the function isVisable() it is assumed that the object is true by default. This is handy when you need an object on the screen all the time even as other objects move off the screen. objects that are not in the viewArea and did not implement isVisable() will still have the render() called on that object this can be a big waste of CPU time so keep the number of objects with out an isVisable() to a minimum amount.