Update Scene.js
added the destroy() function to the scene object so that it can handle closing out anything it may need to prior to the scene being deleted. all objects in the scene are also handled. any objects used to hold values from one scene to another should be put in the game engines direct object array and not in the scenes object array.